// web/src/generated/db.ts
export type Database = {
  public: {
    Tables: {
// ...
      postgrespropertydocuments3: {
        Row: {
          apn: string | null
          block: string | null
          booknumber: string | null
          booktype: string | null
          createdat: string | null
          documentdate: string | null
          documentid: string | null
          filingcode: string | null
          grantee: string | null
          grantor: string | null
          id: number
          latitude: number | null
          longitude: number | null
          lot: string | null
          nameinternalid: string | null
          names: string | null
          numberofpages: number | null
          primarydocnumber: string | null
          secondarydocnumber: string | null
          totalnamescount: number | null
          updatedat: string | null
        }
        Insert: {
          apn?: string | null
          block?: string | null
          booknumber?: string | null
          booktype?: string | null
          createdat?: string | null
          documentdate?: string | null
          documentid?: string | null
          filingcode?: string | null
          grantee?: string | null
          grantor?: string | null
          id?: number
          latitude?: number | null
          longitude?: number | null
          lot?: string | null
          nameinternalid?: string | null
          names?: string | null
          numberofpages?: number | null
          primarydocnumber?: string | null
          secondarydocnumber?: string | null
          totalnamescount?: number | null
          updatedat?: string | null
        }
        Update: {
          apn?: string | null
          block?: string | null
          booknumber?: string | null
          booktype?: string | null
          createdat?: string | null
          documentdate?: string | null
          documentid?: string | null
          filingcode?: string | null
          grantee?: string | null
          grantor?: string | null
          id?: number
          latitude?: number | null
          longitude?: number | null
          lot?: string | null
          nameinternalid?: string | null
          names?: string | null
          numberofpages?: number | null
          primarydocnumber?: string | null
          secondarydocnumber?: string | null
          totalnamescount?: number | null
          updatedat?: string | null
        }
        Relationships: []
      }
      postgrespropertydocuments4: {
        Row: {
          apn: string | null
          block: string | null
          booknumber: string | null
          booktype: string | null
          createdat: string | null
          documentdate: string | null
          documentid: string | null
          filingcode: string | null
          grantee: string | null
          grantor: string | null
          id: number
          latitude: number | null
          longitude: number | null
          lot: string | null
          nameinternalid: string | null
          names: string | null
          numberofpages: number | null
          primarydocnumber: string | null
          secondarydocnumber: string | null
          totalnamescount: number | null
          updatedat: string | null
        }
        Insert: {
          apn?: string | null
          block?: string | null
          booknumber?: string | null
          booktype?: string | null
          createdat?: string | null
          documentdate?: string | null
          documentid?: string | null
          filingcode?: string | null
          grantee?: string | null
          grantor?: string | null
          id?: number
          latitude?: number | null
          longitude?: number | null
          lot?: string | null
          nameinternalid?: string | null
          names?: string | null
          numberofpages?: number | null
          primarydocnumber?: string | null
          secondarydocnumber?: string | null
          totalnamescount?: number | null
          updatedat?: string | null
        }
        Update: {
          apn?: string | null
          block?: string | null
          booknumber?: string | null
          booktype?: string | null
          createdat?: string | null
          documentdate?: string | null
          documentid?: string | null
          filingcode?: string | null
          grantee?: string | null
          grantor?: string | null
          id?: number
          latitude?: number | null
          longitude?: number | null
          lot?: string | null
          nameinternalid?: string | null
          names?: string | null
          numberofpages?: number | null
          primarydocnumber?: string | null
          secondarydocnumber?: string | null
          totalnamescount?: number | null
          updatedat?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}